feature/vocab_tree - RecognitionIO passes tests

+++ main/boofcv-recognition/src/test/java/boofcv/struct/kmeans/TestPackedTupleArray_F64.java
@@ -33,7 +33,7 @@ class TestPackedTupleArray_F64 extends GenericPackedArrayChecks<TupleDesc_F64> {
 
 	@Override protected TupleDesc_F64 createRandomPoint() {
 		var point = new TupleDesc_F64(DOF);
-		point.value[0] = rand.nextGaussian();
+		point.value[0] = (double) rand.nextGaussian();
 		return point;
 	}
